image: the default trait tables reconstruct, and a native-fn crosses by name
The dumper refused every non-nil `traits` field, so no value a running program
allocates could be dumped at all: every collection carries a pointer into its
instance's default tables. A traits slot now has the three answers the design
argues for. Nil crosses as nil. A pointer into the dumping instance's default
table becomes a reconstruction entry, and hydration writes the HYDRATING
instance's table for that tag — a lookup, since VM init built it before any
image could load. Anything else is program data and copies into the body like
the struct it is.

The identity test runs over the whole default table rather than the object's
own tag, because one traitset serves seven tags and `with-traits` can attach
any table to any value. The entry names the first tag sharing the table it
matched; every tag sharing that table resolves to the same one at load.

A reconstructed value comes from the hydrating instance, so neither word of
the slot is known at dump time and the entry rewrites both. Hydration resolves
every constructor before it installs the region, so an instance with no trait
tables refuses the load without leaving a region behind, then increfs and
records the cross-region edge each write creates — the free-time oracle
compares the recorded table against these same pointers.

A traitset's methods are native-fns, which are `Value{TAG_NATIVE_FN, prim_id}`
immediates over an index the canonical tables seed in source order and every
later def extends at run time. The primitive stream carries `(slot, name
index)` against a sorted table, exactly as the file stream does; the dumper
zeroes the payload word, so the artifact records no id. A native-fn root rides
in the header where no stream reaches it, so its payload is a table index and
its own tag says so. A def the canonical tables do not name fails the dump.

The header grows three counts, so the fingerprint offset moves and the format
version goes to 2. format.rs passed the reading budget with them, so it splits
three ways: the geometry and codecs, the header block, and the section map.
